dekaf: remove redundant and broken/bitrotted integration test file, move couple of non-covered tests over
