+++ crates/dekaf/tests/fixtures/field_selection_flow_document.yaml
@@ -1,0 +1,38 @@
+collections:
+  test/collection:
+    key:
+      - /key
+    schema:
+      properties:
+        key:
+          type: string
+        field_a:
+          type: string
+        field_b:
+          type: string
+      type: object
+      required:
+        - key
+        - field_a
+        - field_b
+materializations:
+  test/materialization:
+    endpoint:
+      dekaf:
+        variant: foo
+        config:
+          deletions: kafka
+          token: "1234"
+          strict_topic_names: false
+    bindings:
+      - source: test/collection
+        resource:
+          topic_name: foo
+        fields:
+          include:
+            key: {}
+            flow_document: {}
+          exclude:
+            - field_a
+            - field_b
+          recommended: true
