+++ crates/dekaf/tests/fixtures/fields_not_required.yaml
@@ -1,0 +1,28 @@
+collections:
+  test/collection:
+    key:
+      - /key
+    schema:
+      properties:
+        key:
+          type: string
+        field_a:
+          type: string
+      type: object
+      required:
+        - key
+materializations:
+  test/materialization:
+    endpoint:
+      dekaf:
+        variant: foo
+        config:
+          deletions: kafka
+          token: "1234"
+          strict_topic_names: false
+    bindings:
+      - source: test/collection
+        resource:
+          topic_name: foo
+        fields:
+          recommended: true
